Move the Consul Connect DNS resolution stuff to its own crate.
