+++ src/rust/consul-connect/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod client_dns_config;
+pub mod resolver;
