+++ src/rust/plugin-sdk/generator-sdk/Cargo.toml
@@ -5,6 +5,7 @@ edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 [dependencies]
+consul-connect = { path = "../../consul-connect" }
 rust-proto = { path = "../../rust-proto" }
 thiserror = "1.0.30"
 tokio = { version = "1.17", features = ["full"] }
@@ -13,7 +14,6 @@ tonic-health = "0.5.0"
 tracing = "0.1.29"
 uuid = { version = "0.8", features = ["v4"] }
 serde = { version = "1.0.136", features = ["derive"] }
-trust-dns-resolver = "0.20.3"
 moka = { version = "0.7", features = ["future"] }
 structopt = { version = "0.3", default-features = false }
